requirements.txt

Django>=4.2
django-environ>=0.11
djangorestframework>=3.15
django-filter>=24.2
drf-spectacular>=0.27
whitenoise>=6.7
waitress>=3.0
psutil>=6.0
pystray>=0.19
Pillow>=10.0
python-dotenv>=1.0
